+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Allow posts with attachment and no text.
 - Add `updated` property to edited `Note` objects.
 - Implement `/api/v1/statuses/{status_id}/source` API endpoint.
+- Add API endpoint for updating posts.
 
 ### Changed